+++ src/actions/index.js
@@ -16,6 +16,7 @@ export const SUBMIT_FORM_ACTION = 'SUBMIT_FORM_ACTION';
 export function submitForm(path = '', block_id, data, attachments, captcha) {
   return {
     type: SUBMIT_FORM_ACTION,
+    subrequest: block_id,
     request: {
       op: 'post',
       path: path + '/@submit-form',
@@ -35,12 +36,15 @@ export function submitForm(path = '', block_id, data, attachments, captcha) {
  */
 export const EXPORT_CSV_FORMDATA = 'EXPORT_CSV_FORMDATA';
 
-export function exportCsvFormData(path = '') {
+export function exportCsvFormData(path = '', filename, block_id) {
   return {
     type: EXPORT_CSV_FORMDATA,
+    filename: filename,
     request: {
       op: 'get',
-      path: path + '/@form-data-export',
+      path: `${path}/@form-data-export${
+        block_id ? '?block_id=' + block_id : ''
+      }`,
     },
   };
 }
@@ -51,12 +55,12 @@ export function exportCsvFormData(path = '') {
  */
 export const GET_FORM_DATA = 'GET_FORMDATA';
 
-export function getFormData(path = '') {
+export function getFormData({ path, block_id }) {
   return {
     type: GET_FORM_DATA,
     request: {
       op: 'get',
-      path: path + '/@form-data',
+      path: `${path}/@form-data${block_id ? '?block_id=' + block_id : ''}`,
     },
   };
 }
@@ -67,12 +71,17 @@ export function getFormData(path = '') {
  */
 export const CLEAR_FORM_DATA = 'CLEAR_FORM_DATA';
 
-export function clearFormData(path = '') {
+export function clearFormData({ path, block_id, expired = false }) {
+  const payload = {
+    expired,
+    block_id,
+  };
   return {
     type: CLEAR_FORM_DATA,
     request: {
-      op: 'get',
-      path: path + '/@form-data-clear',
+      op: 'del',
+      path: `${path}/@form-data-clear`,
+      data: payload,
     },
   };
 }