+++ src/components/View.jsx
@@ -8,11 +8,25 @@ import FormView from 'volto-form-block/components/FormView';
 import { formatDate } from '@plone/volto/helpers/Utils/Date';
 import config from '@plone/volto/registry';
 import { Captcha } from 'volto-form-block/components/Widget';
+import { isValidEmail } from 'volto-form-block/helpers/validators';
+import ValidateConfigForm from 'volto-form-block/components/ValidateConfigForm';
 
 const messages = defineMessages({
   formSubmitted: {
     id: 'formSubmitted',
     defaultMessage: 'Form successfully submitted',
+  },
+  defaultInvalidFieldMessage: {
+    id: 'formblock_defaultInvalidFieldMessage',
+    defaultMessage: 'Invalid field value',
+  },
+  requiredFieldMessage: {
+    id: 'formblock_requiredFieldMessage',
+    defaultMessage: 'Fill-in this field',
+  },
+  invalidEmailMessage: {
+    id: 'formblock_invalidEmailMessage',
+    defaultMessage: 'The email is incorrect',
   },
 });
 
@@ -48,12 +62,32 @@ const formStateReducer = (state, action) => {
   }
 };
 
-const getInitialData = (data) => ({
-  ...data.reduce(
-    (acc, field) => ({ ...acc, [getFieldName(field.label, field.id)]: field }),
-    {},
-  ),
-});
+const getInitialData = (data) => {
+  const { static_fields = [], subblocks = [] } = data;
+
+  return {
+    ...subblocks.reduce(
+      (acc, field) =>
+        field.field_type === 'hidden'
+          ? {
+              ...acc,
+              [getFieldName(field.label, field.id)]: {
+                ...field,
+                ...(data[field.id] && { custom_field_id: data[field.id] }),
+              },
+            }
+          : acc,
+      {},
+    ),
+    ...static_fields.reduce(
+      (acc, field) => ({
+        ...acc,
+        [getFieldName(field.label, field.id)]: field,
+      }),
+      {},
+    ),
+  };
+};
 
 /**
  * Form view
@@ -62,22 +96,24 @@ const getInitialData = (data) => ({
 const View = ({ data, id, path }) => {
   const intl = useIntl();
   const dispatch = useDispatch();
-  const { static_fields = [] } = data;
 
   const [formData, setFormData] = useReducer((state, action) => {
     if (action.reset) {
-      return getInitialData(static_fields);
+      return getInitialData(data);
     }
 
     return {
       ...state,
       [action.field]: action.value,
     };
-  }, getInitialData(static_fields));
+  }, getInitialData(data));
 
   const [formState, setFormState] = useReducer(formStateReducer, initialState);
   const [formErrors, setFormErrors] = useState([]);
-  const submitResults = useSelector((state) => state.submitForm);
+
+  const submitResults = useSelector(
+    (state) => state.submitForm?.subrequests?.[id],
+  );
   const captchaToken = useRef();
 
   const onChangeFormData = (field_id, field, value, extras) => {
@@ -108,30 +144,58 @@ const View = ({ data, id, path }) => {
         config.blocks.blocksConfig.form.additionalFields?.filter(
           (f) => f.id === fieldType && f.isValid !== undefined,
         )?.[0] ?? null;
+
       if (
         subblock.required &&
         additionalField &&
         !additionalField?.isValid(formData, name)
       ) {
-        v.push(name);
+        const validation = additionalField?.isValid(formData, name);
+        if (typeof validation === 'boolean') {
+          //for retro-compatibility with previous formErrors structure
+          v.push({
+            field: name,
+            message: intl.formatMessage(messages.defaultInvalidFieldMessage),
+          });
+        } else if (typeof validation === 'object') {
+          v.push(validation);
+        }
       } else if (
         subblock.required &&
         fieldType === 'checkbox' &&
         !formData[name]?.value
       ) {
-        v.push(name);
+        v.push({
+          field: name,
+          message: intl.formatMessage(messages.requiredFieldMessage),
+        });
       } else if (
         subblock.required &&
         (!formData[name] ||
           formData[name]?.value?.length === 0 ||
           JSON.stringify(formData[name]?.value ?? {}) === '{}')
       ) {
-        v.push(name);
+        v.push({
+          field: name,
+          message: intl.formatMessage(messages.requiredFieldMessage),
+        });
+      } else if (
+        fieldType === 'from' &&
+        formData[name]?.value &&
+        !isValidEmail(formData[name].value)
+      ) {
+        v.push({
+          field: name,
+          message: intl.formatMessage(messages.invalidEmailMessage),
+        });
       }
     });
 
     if (data.captcha && !captchaToken.current) {
-      v.push('captcha');
+      v.push({
+        field: 'captcha',
+        message: intl.formatMessage(messages.requiredFieldMessage),
+      });
     }
 
     setFormErrors(v);
@@ -158,7 +222,10 @@ const View = ({ data, id, path }) => {
             let name = getFieldName(subblock.label, subblock.id);
             if (formattedFormData[name]?.value) {
               formattedFormData[name].field_id = subblock.field_id;
-              const isAttachment = subblock.field_type === 'attachment';
+              const isAttachment =
+                config.blocks.blocksConfig.form.attachment_fields.includes(
+                  subblock.field_type,
+                );
               const isDate = subblock.field_type === 'date';
 
               if (isAttachment) {
@@ -205,10 +272,16 @@ const View = ({ data, id, path }) => {
     setFormState({ type: FORM_STATES.normal });
   };
 
+  const getErrorMessage = (field) => {
+    const e = formErrors?.filter((e) => e.field === field);
+    return e.length > 0 ? e[0].message : null;
+  };
+
   const captcha = new Captcha({
     captchaToken,
     captcha: data.captcha,
     captcha_props: data.captcha_props,
+    errorMessage: getErrorMessage('captcha'),
     onChangeFormData,
   });
 
@@ -246,18 +319,21 @@ const View = ({ data, id, path }) => {
   }, []);
 
   return (
-    <FormView
-      id={formid}
-      formState={formState}
-      formErrors={formErrors}
-      formData={formData}
-      captcha={captcha}
-      onChangeFormData={onChangeFormData}
-      data={data}
-      onSubmit={submit}
-      resetFormState={resetFormState}
-      resetFormOnError={resetFormOnError}
-    />
+    <ValidateConfigForm data={data}>
+      <FormView
+        id={formid}
+        formState={formState}
+        formErrors={formErrors}
+        formData={formData}
+        captcha={captcha}
+        onChangeFormData={onChangeFormData}
+        data={data}
+        onSubmit={submit}
+        resetFormState={resetFormState}
+        resetFormOnError={resetFormOnError}
+        getErrorMessage={getErrorMessage}
+      />
+    </ValidateConfigForm>
   );
 };